+++ common/src/main/java/mca/entity/ai/MemoryModuleTypeMCA.java
@@ -23,6 +23,7 @@ public interface MemoryModuleTypeMCA {
     RegistrySupplier<MemoryModuleType<LivingEntity>> NEAREST_GUARD_ENEMY = register("nearest_guard_enemy", Optional.empty());
     RegistrySupplier<MemoryModuleType<Boolean>> WEARS_ARMOR = register("wears_armor", Optional.of(Codec.BOOL));
     RegistrySupplier<MemoryModuleType<Integer>> SMALL_BOUNTY = register("small_bounty", Optional.of(Codec.INT));
+    RegistrySupplier<MemoryModuleType<LivingEntity>> HIT_BY_PLAYER = register("hit_by_player", Optional.empty());
 
     static void bootstrap() {
         MEMORY_MODULES.register();